Modify DTOs to use parameterless constructors + example adapter update
Most DTOs previously used get-only properties and constructors that accepted a parameter for each property. This has been modified so that all DTOs now have a parameterless constructor, (and accompanying Create methods that perform the same function as the old constructors), to allow compatibility with serializers such as System.Text.Json that do not currently support deserialization of classes that do not have default constructors.

Additionally, the example adapter has been updated to support asset model browsing. The intention is to ultimately provide implementations for all features on this adapter.