DataFunctionDescriptor IDs + aggregation helper improvements
DataFunctionDescriptor now has a separate ID, to allow the name to be locale-specific.

AggregationHelper/InterpolationHelper/PlotHelper now allow values for multiple tags to be calculated in a single call (i.e. allow the helper to handle splitting the raw data channel into separate channels for each tag in the request)

Add AggregationHelper implementation of Range and Count functions.