Topic-based subscriptions can be modified after creation again
Snapshot tag value and topic-based event subscriptions are now implemented using bidirectional streaming again to allow for subscription changes after the subscription has been created.

In the gRPC client/server, this is achieved using `oneof` in the request stream messages to distinguish between the initial request to create the stream and subsequent changes to the subscription.