Tag search result fields
Adds the ability to specify a `ResultFields` property on a `FindTagsRequest` to control which parts of a tag definition a caller is interested in receiving. Basic information (name, ID, description, units, data type) is always returned; enum flags can be set to specify if properties, digital states, and labels should also be included.