Allow URI hash when adding a supported feature to a tag definition
In some cases, it may be necessary to specify e.g. exactly which data functions are supported when requesting aggregated data for a tag. TagDefinitionBuilder now allows a hash to be specified when adding supported features to a tag, meaning that a tag can add the same feature multiple times but with a different hash to indicate which features are supported.