Allow tag values to have multiple values
`TagValue` now includes an `AdditionalValues` property that can be used to define additional secondary values for a sample. This can be used when e.g. the primary `Value` property holds the numerical value of a digital state, but you also want to return the name of the state for display purposes.