Revert to single value on TagValue objects
The move to allow multiple values on `TagValue` and `TagValueExtended` types in #93 is unnecessarily confusing, given that the only real use case is to be able to return both the numeric value and display value for a digital state.

This commit reverses the changes by removing the `Values` property on a tag value and re-adding the original `Value` property, so that the value is only a single `Variant` again.

To allow digital state names to be passed as part of a `TagValueExtended`, the `WellKnownProperties` type has been added to provide well-known property names that can be added to adapter entities. The display value for a tag value can therefore be specified as a bespoke property on the tag value.

There is still a use case for allowing multiple values to be passed in a tag value (e.g. the use of multidimensional arrays in OPC UA, or data sets/data tables in MTConnect), but this is best solved by improvements to the `Variant` type itself to support arrays (and possibly key-value maps).