Object encoder signature changes
`IObjectEncoder` has been pared back to focus on conversion of objects to/from `byte[]`, with extension methods providing the ability to directly encode to/from `EncodedObject`.

This approach allows for better wrapping of encoders (e.g. to allow the object payload to be encrypted by a wrapper `IObjectEncoder` after serialization).