Add ExtensionFeatureDataTypeAttribute.cs
`ExtensionFeatureDataTypeAttribute` extends `DataTypeIdAttribute` to simplify registration of type IDs for types that are strongly tied to an extension feature. The attribute allows the consumer to specify a relative name for the type, which is then made absolute using the URI if the extension feature e.g. `ping-message` => `asc:extensions/example/ping-pong/types/ping-message/`