Miscellaneous updates (#112)
* Update JsonObjectEncoder.cs

`Default` now returns an instance of `IObjectEncoder`

* Update example adapter

Update extension feature ping method to demonstrate how to return multiple variant values

* Update SnapshotTagValuePush.cs

Allow `SnapshotTagValuePush` to keep cached tag values even when a tag has no subscribers, and add a `ReadSnapshotTagValues` method that is compatible with `IReadSnapshotTagValues`.

This allows a `SnapshotTagValuePush` instance to be used as a cache when an adapter implements `IReadSnapshotTagValues` if required.

* Ensure that JsonStringEnumConverter is registered

- Add `AddDataCoreAdapterConverters` extension method for `JsonSerializerOptions`.
- Ensure that a `JsonStringEnumConverter` is registered in the JSON serializer options.

* Update AdapterRequestExtensions.cs

Add new extension for selection a page of ordered items using an explicit page size and page number

* Update AssetModelNodeExtensions.cs

Add extension method for automatically applying a `BrowseAssetModelNodesRequest` if you have a list of all asset model nodes available (e.g. if the nodes are stored in-memory).

* Update TagDefinitionExtensions.cs

Update `ApplyFilter` extension method so that it will also apply the `ResultFields` filter of a `FindTagsRequest` request.