Simplify IAdapterAccessor and AdapterAccessor implementation
`IAdapterAccessor` no longer makes an explicit demand on whether only enabled adapters should be returned or not; the expected behaviour is that only enabled adapters will be returned.

Additionally, when processing a `FindAdapters` request, `AdapterAccessor` no longer retrieves all adapters from the implementing class and applies filtering and paging in the base class. Instead, it provides helper methods to allow the implementing class to determine if an adapter matches a filter or if the caller is authorized to access the adapter.