Allow use of bespoke interceptors in gRPC adapter proxy
Adds a new  `GetClientInterceptors` property to `GrpcAdapterProxyOptions`, which can be used to provide gRPC interceptors to attach to all gRPC client objects created by a `GrpcAdapterProxy`.

The motivation is to allow OpenTelemetry support to be added to gRPC adapter proxies that are using Grpc.Core instead of Grpc.Net.Client (via [this package](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/tree/main/src/OpenTelemetry.Contrib.Instrumentation.GrpcCore)).

This also allows arbitrary interceptors to be added, which may be of general use anyway.