Update CommonAdapterConfigurationExtensions.cs
Implements #138.

When calling `AddHostInfo` and no vendor information is supplied, the method will now register a factory method that will check the service provider for a `VendorInfo` service at resolution time. If no `VendorInfo` is registered with the service provider, it will look for a `[VendorInfo]` annotation on the assembly returned by `Assembly.GetEntryAssembly()`.