Add RFC 7807 support to adapter HTTP client
Adds RFC 7807 problem details support to the `AdapterHttpClient`.

When a non-good status code is returned by a call to an adapter API endpoint, an `AdapterHttpClientException` (derived from `HttpRequestException`) is thrown. If the response contains a JSON-encoded problem details object (`application/problem+json`), this is deserialized and assigned to the `ProblemDetails` property on the exception.

Although the client already ensures that e.g. request payloads are validated before sending, being able to pass through problem details responses is clearly useful when it comes to extracting useful information about server-side errors.