Update CertificateUtilities.cs
Remove `TryLoadCertificateFromStore` overload that allowed a `validOnly` flag. The behaviour is now that invalid certificates can also be returned, but a new `SelectCertificate` method will select an appropriate certificate from the collection returned from the store.

The selection methodology used when more than one certificate is returned from the store is:

- If any of the certificates are valid using basic validation policy, the matching certificate with the latest NotAfter value is returned.
- If any of the certificates are valid using basic validation policy but allowing unknown root certificates (e.g. valid self-signed certificates), the matching certificate with the lasted NotAfter value is returned.
- If no certificates are in the collection that meet the above criteria, the certificate with the latest NotAfter value is returned (bearing in mind that this certificate will fail default validation tests!)