Use in-memory index
File system store now uses an in-memory lookup from key to file. Files are assigned to different folders based on the hash code of the key.

The index is persisted to disk on change and loaded when the store is first queried.

The index is always gzipped (using fastest possible compression); data files can be optionally gzipped via FileSystemKeyValueStoreOptions.