Adapter can start in response to options change
In `AdapterBase<T>`, an adapter that is created using a, `IOptionsMonitor<T>` will attempt to start itself on options change if it is not already running if `StartAsync` has been called at least once on the adapter.

Calling `StartAsync` on a disabled adapter no longer throws an exception. Instead, a warning message is logged and `StartAsync` will exit immediately.