Rename templated adapter class
Templated adapter class has changed from `ExampleHostedAdapter` to `ExampleHostedAdapterImpl`. This means that we no longer end up with an adapter class with the same name as its namespace.

The OpenTelemetry service ID for the host is also now set to the adapter ID by default, rather than the DNS name of the machine.