Modify templated class names
Template adapter and options classes have been renamed to `MyAdapter` and `MyAdapterOptions` respectively.

This is to allow an out-of-the-box project to work correctly when someone specifies a project name such as `MyCompany.SomeAdapter`, as this was previously causing the adapter class to be renamed to be the same as the project name with an `Impl` suffix, which led to compilation errors when trying to run the new project.