Add default HTTP request version
Adds a `DefaultRequestVersion` property that gets used (when non-null) when creating `HttpRequestMessage` instances in the various non-static `CreateHttpRequestMessage` overloads.

The static `CreateHttpRequestMessage` overloads have all been modified to accept a `Version?` parameter.

The `CreateHttpVersionHandler` method for creating a `DelegatingHandler` that sets the HTTP version of the request has been removed. Any requests that have content set before the request version will automatically HTTP 1.1, as HTTP 1.1 is the default `HttpRequestMessage` version, and setting request content when the version is 1.1 causes a `Transfer-Encoding: chunked` header to be added to the request.