Don't use JsonSerializerContext
Removes the creation of a JsonSerializerContext. There are breaking changes in .NET 7 around how serialization works when using a JsonSerializerContext (or more specifically, when the serializer encounters a type not included in the context).

Better to revisit the use of the serializer context at a later date when we can look more closely at the performance benefits versus inflexibility trade-off.