Allow adapters to skip request validation (#276)
* Add call context helpers

Adds general Set/TryGet helpers for entries in the `IAdapterCallContext.Items` dictionary.

Also adds helpers for getting/setting a flag in `IAdapterCallContext.Items` that specifies if `AdapterFeatureWrapper<TFeature>` should skip request validation.

* Allow validation to be skipped

Validation will be skipped in `AdapterFeatureWrapper<TFeature>` based on the new `UseRequestValidation` extension method on `IAdapterCallContext`.

* Allow registration of a wrapped feature

When registering a feature with `AdapterCore`, allow the feature to be an instance of `AdapterFeatureWrapper`.

* Make abstractions internals visible to unit tests and allow unit tests to register custom "standard" adapter features for test purposes

* Enable adapter by default

Default behaviour in `AdapterCore` is now to enable the adapter by default. Behaviour in `AdapterBase<T>` has been inverted so that it now disables the adapter if the adapter options specify that it should be disabled.

* Skip request validation for REST endpoints

Adds an action filter to adapter MVC controllers that adds an item into the HttpContext's Items collection (which are directly used by the HttpAdapterCallContext) to disable or enable adapter-level validation.