Fixes minimal API validation issues with URI properties
The MiniValidation NuGet package is used for validating requests in minimal API routes but had issues in versions 0.7.2 and 0.7.3 that meant that requests containing `Uri` request properties would cause exceptions when validated.

The workaround was to have a temporary request types for models that replaced `Uri` properties with `string`s instead, that were then mapped to the original model types.

This issue has been fixed in MiniValidation 0.7.4.