Fix warnings when packing templates project
When packing the DataCore.Adapter.Templates project, two NuGet-related build warnings are raised:

- NU5119: raised because the project that template files are copied from includes a .gitignore file
- NU5128: raised because the templates project file targets .NET Standard 2.0, but uses the `IncludeBuildOutput` property to prevent the project output from included in the package (since the templates package only contains a `content` folder).

CI build is now configured to treat all build warnings as errors which results in the pack task failing because of the above warnings.

The fix adds a new exclusion for template source files that start with `.` and sets the `SuppressDependenciesWhenPacking` build property to `true`.