Handle undefined values in JsonElementConverter
Updates `JsonElementConverter` to serialize instances with a `ValueKind` property of `System.Text.Json.JsonValueKind.Undefined` as `null`.

This is not an ideal approach, but undefined values currently get serialized as an empty token (which produces invalid JSON) and using `JsonWriter.WriteUndefined` writes a token of `undefined`, which is also invalid JSON.

Writing a `null` value means that the `JsonElement` and its serialized version are no longer equivalent, but it means that we produce valid JSON which is a more important factor.