Add interpolated query start/end values
Updates PlotHelper so that values are interpolated exactly at the query start and end time.

Also modifies default behaviour for non-numeric tags so that a maximum of 6 samples per bucket are returned. In the previous implementation, every change in value or quality for a non-numeric tag was returned, meaning that it was possible to return every sample in the bucket no matter how many there were.