Add ToEnumerable overload with expected item count
Adds a new `ChannelExtensions.ToEnumerable` overload that accepts an `expectedItems` parameter.

The `expectedItems` parameter is used to configure the initial capacity of the `List<T>` that items are added to instead of the `maxItems` parameter. This is to allow for scenarios where the maximum number of items allowed to be read from the `IAsyncEnumerable<T>` is high but the expected number of items is much lower. For example, a tag data query might be allowed to return up to 20 000 samples, but the average number of samples returned by a query might only be 1000. It therefore makes more sense to initialise the list with a capacity of 1000 rather than 20 000.