Hosting improvements (#348)
* Use scoped services instead of singletons

Modifies `CommonAdapterConfigurationExtensions` to register the `IAdapterAuthorizationService` as a scoped service instead of a singleton.

This change is to ensure that `DefaultAdapterAuthorizationService` can be correctly constructed at runtime. `DefaultAdapterAuthorizationService` depends on ASP.NET Core's `IAuthorizationService`. `IAuthorizationService` is registered as transient, but becomes scoped if any `IAuthorizationHandler` services or their dependencies are registered as scoped services.

Changing the registration of `IAdapterAuthorizationService` to scoped has no impact on the processing of HTTP requests, but does require an update to the `AdapterInitializer` hosted service to create a new scope before attempting to resolve adapters at startup time.

Additionally, `IAdapterAccessor` is now also registered as a scoped service.

* Remove OpenAPI/API versioning

Removes OpenAPI and API versioning from Minimal APIs project for the time being.

Both of these features will be revisited in the future.