Refactor subscription management
The current subscription management base classes add a lot of unnecessary cruft to subscription management and don't perform as well as they could when it comes to matching message topics to subscribers.

This commit uses MQTTnet as its basis/inspiration for more lightweight subscriptions that can optionally support MQTT-style wildcards. There is still some work required (such as allowing the level separators and single-multi-level wildcards to be customised), but it contains working implementations of `ISnapshotTagValuePush` helpers (including one that polls for new values under the hood).