Use background database flush if configured
Updates `SqliteKeyValueStore` to allow a background flush interval to be specified (opt-in; default behaviour is to continue to write to the database immediately).

When a flush interval is specified, write and delete operations are put into a dictionary indexed by data store key (value = the bytes to write for that key; null value = delete the key). A background task performs a flush at the interval specified.