Add key/value store write buffer helper
`KeyValueStoreWriteBuffer` is a new helper class that replicates (and extends) the write buffering in the SQLite key/value store so that the buffering can be used in any implementation that requires it.

In addition to periodic flushing of pending changes, `KeyValueStoreWriteBuffer` also allows automatic flushing to be configured when the number of pending changes exceeds a limit or the total byte size or the pending changes exceeds a limit.

The SQLite-based store has been rewritten to use `KeyValueStoreWriteBuffer` instead of its own implementation.