Ensure that byte[] is always converted to ByteString (#372)
* Update ByteString

Adds a new constructor that accepts a base64-encoded byte sequence.

Adds a `TryParse` method for creating from a base64-encoded byte sequence.

Adds equality operator overrides.

* Variant always treats byte[] as ByteString

When creating a new `Variant` using a `byte[]` value, the variant will now always convert the value to a `ByteString` internally and will set the variant type to be `VariantType.ByteString`. The `ArrayDimensions` property for a variant created using `byte[]` will now always be `null`.

`TryGetVariantType` no longer checks the element type if the object passed is an array i.e. the result will now be false for all array types except `byte[]`.

Special handling has been added in various places to ensure that `byte[]` is always converted to `ByteString` in relevant constructors. The exception to this rule is in the internal constructor used by the JSON converter.

* Update PublicAPI.Unshipped.txt

* Add legacy byte[] variant deserialize test