Add scoped logger/logger factory
`ScopedLogger` is a disposable `ILogger` that wraps an inner `ILogger` and creates a logging scope that is disposed when the logger is disposed.

The intention here is to allow `AdapterCore` to create a scoped logger when it is constructed that is disposed when the adapter is disposed. This ensures that the scope flows into any helper utilities that are constructed using the adapter's logger.