Add logger factory
Adds a `LoggerFactory` property to `AdapterCore` to enable adapters to create loggers for helper types etc.

A new constructor has been added to allow an `ILoggerFactory` to be provided. If the existing constructor that accepts `ILogger` is used, a shim logger factory that always returns this logger will be created.