Scoped loggers (#376)
* Add scoped logger/logger factory

`ScopedLogger` is a disposable `ILogger` that wraps an inner `ILogger` and creates a logging scope that is disposed when the logger is disposed.

The intention here is to allow `AdapterCore` to create a scoped logger when it is constructed that is disposed when the adapter is disposed. This ensures that the scope flows into any helper utilities that are constructed using the adapter's logger.

* Use source-generated log methods

* Set package versions in templates with PowerShell

Updates the templates project to use a PowerShell script to set package versions in the final templated project files.

This allows us to automatically pull the correct versions from Directory.Packages.props instead of having to reference specific build properties for e.g. OpenTelemetry package versions.

* Use Serilog in example adapter host

Updates the example adapter host to use Serilog for logging. This makes it much easier to add non-console logging destinations to out-of-the-box adapter hosts.

* Add logger factory property

Adds a `LoggerFactory` property to `AdapterCore` to enable adapters to create loggers for helper types etc.

A new constructor has been added to allow an `ILoggerFactory` to be provided. If the existing constructor that accepts `ILogger` is used, a shim logger factory that always returns this logger will be created.

* Deprecate constructors that accept ILogger

* Deprecate Logger property

* Add AdapterBase<T> overloads that accept ILoggerFactory

* Require typed loggers in helper classes

* Update docs

* Update shipped API details