Tag value builder properties are now distinct
`TagValueBuilder` now enforces that only one property with a given name can be created on a tag value. If property with the same name already exists when a property is set on the builder, the existing property is replaced.

Various property-related methods have been moved to `TagValueBuilderExtensions` as extension methods. A new extension method for setting the value to a digital state has also been added; this will set both a numerical value and a display value.