Add UTC DateTime converters
Adds System.Text.Json converters for DateTime and DateTime? that ensure that values are always converted to UTC prior to serialization, and are always converted to UTC on deserialization.

This ensures that sample times, event times, etc. will always be in UTC in query results or when writing to an adapter via the HTTP or SignalR API. gRPC already explicitly uses UTC when converting from the protobuf Timestamp type to DateTime and vice versa.