Enforce UTC conversion when serializing DateTime instances (#394)
* Add UTC DateTime converters

Adds System.Text.Json converters for DateTime and DateTime? that ensure that values are always converted to UTC prior to serialization, and are always converted to UTC on deserialization.

This ensures that sample times, event times, etc. will always be in UTC in query results or when writing to an adapter via the HTTP or SignalR API. gRPC already explicitly uses UTC when converting from the protobuf Timestamp type to DateTime and vice versa.

* Ensure that tag value sample time is UTC

Ensures that TagValue.UtcSampleTime is always converted to UTC on creation.

* Ensure that gRPC proxy uses UTC timestamps

Protobuf Timestamp.FromDateTime throws an exception if a non-UTC DateTime is provided. This commit updates the gRPC proxy to ensure that it converts query time ranges to UTC when constructing the equivalent gRPC request object.

* Enforce UTC conversion for Newtonsoft

Modifies JsonSerializerSettingsExtensions to ensure that JsonSerializerSettings objects always have their DateTimeZoneHandling property set to DateTimeZoneHandling.Utc, and adds an IsoDateTimeConverter that always adjusts timestamps to UTC.

* Add DateTime JSON conversion tests

Adds tests for System.Text.Json and Newtonsoft to ensure that DateTime instances are converted to UTC on serialize and deserialize.

* Update version.json

* Update PublicAPI.Shipped.txt