Use random Kestrel port number for unit tests
The port number to use for Kestrel when running unit tests is now a random number in the range [31415,31500) instead of a constant value of 31415.

This ensures that, if unit tests target multiple frameworks and tests for each framework are run in parallel, there is less chance of a clash occurring.