Add FeatureBase
FeatureBase is a new class that can serve as the base type for feature implementations that are external to an IAdapter implementation.

FeatureBase provides built-in support for health checks and can be easily extended to provide inner health checks and/or custom health check data to add to the composite top-level health check result.