Add base class for external feature implementations (#421)
* Remove unsupported frameworks

Remove all net6.0 and net7.0 targets, as these are now out of support

* Package updates

* Update version

* Update DataCore.Adapter.Tests.csproj

* Add FeatureBase

FeatureBase is a new class that can serve as the base type for feature implementations that are external to an IAdapter implementation.

FeatureBase provides built-in support for health checks and can be easily extended to provide inner health checks and/or custom health check data to add to the composite top-level health check result.

* Inherit from FeatureBase

Modifies various feature implementations to extend from FeatureBase

* Update public API definitions