Better feature unwrapping logic
Modifies the `AdapterFeatureExtensions.Unwrap<T>` extension method so that no longer checks to see if the feature is an `AdapterFeatureWrapper<T>` and instead checks to see if it is an `AdapterFeatureWrapper` with an `InnerFeature` property that is an instance of the provided `T`.

This allows `Unwrap<T>` to succeed when `T` is `IAdapterFeature` rather than a more-derived type.