AdapterBase<TAdapterOptions> supports custom functions by default (#429)
* Include custom function ID in authz handler

Breaking change. Modifies the signature for CustomFunctionAuthorizeHandler to include the ID of the custom function that is being authorized.

* Allow default custom function authz handler

Modifies CustomFunctions implementation to allow a default authorization handler to be specified if an individual custom function does not supply its own authorization handler.

Also allows the JsonSerializerOptions for the CustomFunctions instance to be modified after creation.

* AdapterBase<T> implements custom functions

Adds a CustomFunctions property to AdapterBase<T> that can be used to register custom functions by default.

* Update PublicAPI.Unshipped.txt

* Use inherited custom functions manager

* Proxies don't use default custom functions manager

* Update writing-an-adapter.md

* All constructors set custom functions field

* Use inherited custom functions manager

* Add synchronous register/unregister methods

* Update PublicAPI registrations